BRONZIE/ai_employee/pyproject.toml
[project]
name = "ai-employee-bronze"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "schedule>=1.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
